fix some more
