+++ src/aioway/hop/hop.py
@@ -21,7 +21,7 @@ from aioway._utils import (
 )
 from aioway.attrs import Attr, AttrDict
 
-__all__ = ["Hop", "TensorHop", "TdictHop", "ListHop", "hop_dcls"]
+__all__ = ["hop_dcls", "Hop", "TensorHop", "TdictHop", "ListHop", "BoundedHop"]
 
 
 @typing.dataclass_transform()
@@ -237,3 +237,30 @@ class ListHop(Hop[cabc.Sequence[typing.Any]]):
 
         dag_nodes = [DagNode(key=hop, deps=list(hop.deps())) for hop in self.nodes()]
         return topo_sort(dag_nodes)
+
+
+@hop_dcls
+class BoundedHop(TdictHop, abc.ABC):
+    """
+    A stream with `__len__` and `__getitem__`.
+    """
+
+    @abc.abstractmethod
+    def __len__(self) -> int:
+        "The number of batches saved in the current `Stream`."
+
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def __getitem__(self, key: int, /) -> td.TensorDict:
+        """
+        Get individual items. Does not support slice input.
+
+        Args:
+            idx: An integer. Must be in the range `[-len(self), len(self))`.
+
+        Returns:
+            The `td.TensorDict` batch.
+        """
+
+        raise NotImplementedError
